use ApiTable for employee table
